backend/requirements.txt
fastapi
uvicorn
pydantic
requests
openai
python-dotenv
web3>=6.10.0
